+++ tests/unit_tests/cx_core/integration/mqtt_test.py
@@ -1,6 +1,7 @@
 from typing import Dict
 
 import pytest
+from appdaemon.plugins.mqtt.mqttapi import Mqtt
 from cx_core.controller import Controller
 from cx_core.integration.mqtt import MQTTIntegration
 from pytest_mock.plugin import MockerFixture
@@ -34,3 +35,22 @@ async def test_callback(
         handle_action_patch.assert_called_once_with(expected)
     else:
         handle_action_patch.assert_not_called()
+
+
+@pytest.mark.asyncio
+async def test_listen_changes(
+    fake_controller: Controller,
+    mocker: MockerFixture,
+):
+    controller_id = "controller_id"
+    listen_event_mock = mocker.patch.object(Mqtt, "listen_event")
+    mqtt_integration = MQTTIntegration(fake_controller, {})
+
+    await mqtt_integration.listen_changes(controller_id)
+
+    listen_event_mock.assert_called_once_with(
+        fake_controller,
+        mqtt_integration.event_callback,
+        topic=controller_id,
+        namespace="mqtt",
+    )