+++ tests/unit_tests/cx_core/type/cover_controller_test.py
@@ -8,7 +8,7 @@ from cx_core.feature_support.cover import CoverSupport
 from cx_core.type_controller import TypeController
 from pytest_mock.plugin import MockerFixture
 
-from tests.test_utils import fake_fn, wrap_exetuction
+from tests.test_utils import fake_fn, wrap_execution
 
 ENTITY_NAME = "cover.test"
 
@@ -55,7 +55,7 @@ async def test_init(
         "close_position": close_position,
     }
 
-    with wrap_exetuction(error_expected=error_expected, exception=ValueError):
+    with wrap_execution(error_expected=error_expected, exception=ValueError):
         await sut_before_init.init()
 
     if not error_expected: